backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
pandas==2.2.3
httpx==0.28.1
python-multipart==0.0.18
python-dateutil==2.9.0.post0
openai==1.58.1
requests==2.32.3
rich==13.9.4
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt<5.0.0
authlib==1.3.2
email-validator==2.3.0
psycopg2-binary==2.9.9
